shell: fix waiting for the child processes
Our signal handler for SIGCHLD may run any time reaping any
terminated child processes.
It even may reap processes the main loop is explicitly waiting for.
This may cause the status of some processes not being reported properly.

Imagine simply calling a non-exiting binary.
The forked child will exit with status 1 after execvp failed.
The signal handler of the shell may run before the specific "blocking"
waitpid call is reached, reaping the child and dropping its exit status.

In order to prevent this race, we block SIGCHLD before forking.
We immediately enable it again in the forked child.

In the shell however, the signal is unblocked after we have waited for
the child or immediately if we are not blocking.