chown: use a saveptr for strtok
Since strtok is used in the getpwnam call we must provide our own
saveptr to not lose state during the user and group extraction.

Since parsing the passwd file during the getipwnam call the
last string buffer of strtok will no longer poihnt to our argv[1].
We have to manually find the start of the group.